rename to WeightCommand
